% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_mbc.R
\name{boxplot.mbc}
\alias{boxplot.mbc}
\title{Boxplot of \code{mbc} timings.}
\usage{
\method{boxplot}{mbc}(x, unit = "t", log = TRUE, xlab, ylab, ...)
}
\arguments{
\item{x}{A \code{mbc} object.}

\item{unit}{Unit in which the results be plotted.}

\item{log}{Should times be plotted on log scale?}

\item{xlab}{X axes label.}

\item{ylab}{Y axes label.}

\item{...}{Passed on to boxplot.formula.}
}
\description{
Boxplot of \code{mbc} timings.
}
\examples{
m1 <- mbc(sum(rexp(ceiling(100*rexp(1)))))
boxplot(m1)

}
\author{
Collin Erickson, Olaf Mersmann
}
