% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-sampler.R
\name{estim_accuracy}
\alias{estim_accuracy}
\title{Estimate the classification accuracy with MCMC estimate of membership probability and ground truth membership.}
\usage{
estim_accuracy(prob, true_z, order)
}
\arguments{
\item{prob}{MCMC samples of \eqn{\Lambda} of dimension `N x K` from [mbc::gibbs_sampler()].}

\item{true_z}{Ground truth \eqn{z} of dimension `N`.}

\item{order}{Optional true order of the membership indicator with length `K` obtained from [mbc::relabel()].}
}
\value{
A list consisting `prob`, the probability of each observation being successfully classified with dimension `N x 2`,
        and `expectation` the overall probability of successfully classifying with length `2`.
}
\description{
Estimate the classification accuracy with MCMC estimate of membership probability and ground truth membership.
}
