% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{freq_table}
\alias{freq_table}
\title{Count the number of observations within each cluster.}
\usage{
freq_table(z, K)
}
\arguments{
\item{z}{Cluster membership indicator vector of length `N`.}

\item{K}{Number of clusters (positive integer).}
}
\value{
A named vector of length K consisting of number of observations (`N_k`) in each cluster.
}
\description{
Count the number of observations within each cluster.
}
