% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-init.R
\name{init_params}
\alias{init_params}
\title{Parameter initialization routine.}
\usage{
init_params(y, K, type = "kmeans", max_iter = 10, nstart = 10)
}
\arguments{
\item{y}{Observed data of dimension `N x q`.}

\item{K}{Number of clusters (positive integer).}

\item{type}{The type of initialization method. Can be either "kmeans" or "random".}

\item{max_iter}{The maximum number of steps in the [stats::kmeans()] algorithm.}

\item{nstart}{The number of random starts in the [stats::kmeans()] algorithm.}
}
\value{
A list consisting of:
z0: initial estimate of the cluster membership indicator vector of length `N`,
\eqn{\rho}: initial estimate of Gaussian Mixture model weight of vector of length `K`,
\eqn{\Theta}: inital estimate of unoberved data of dimension `N x q`,
\eqn{\mu}: initial estimate of cluster means matrix of dimension `K x q`.
\eqn{\Sigma}: initial estimate of cluster variance array of dimension `q x q x K`.
}
\description{
Parameter initialization routine.
}
\details{
This functions uitlizes \code{init_z} from [nnm-functions].
}
\examples{
# This is a simulated example
N <- 100
q <- 5
K <- 4
true_z <- sample(1:K, N, TRUE)
true_mu <- matrix(rnorm(q*K, 0, 5), ncol = q)
true_Sigma <- replicate(K, expr = {mniw::rwish(1, diag(q)*runif(1, 0.01, 0.1), q+1)})
true_Theta <- mniw::rmNorm(N, true_mu[true_z,], true_Sigma[,,true_z])
V <- replicate(N, expr = {mniw::rwish(1, diag(q)*0.05, q+1)})
y <- mniw::rmNorm(N, true_Theta, V)
init_params(y, K, type = "kmeans")

}
