% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{ll}
\alias{ll}
\title{Complete log-likelihood for the normal-normal-mixture model.}
\usage{
ll(y, Theta, V, Sigma, mu, z, rho, N, q, K, v_k, omega)
}
\arguments{
\item{y}{Observed data of dimension `N x q`.}

\item{Theta}{Unobserved data matrix of dimension `N x q`.}

\item{V}{Observed Fisher information array of dimension `q x q x N`.}

\item{Sigma}{Cluster variance array of dimension `q x q x K`.}

\item{mu}{Cluster means matrix of dimension `K x q`.}

\item{z}{Cluster membership indicator vector of length `N`.}

\item{rho}{Gaussian Mixture model weight of vector of length `K`.}

\item{N}{Number of observations (positive integer).}

\item{q}{Size of \eqn{\theta_i} (positive integer).}

\item{K}{Number of clusters (positive integer).}

\item{v_k}{Baseline degree of freedom for Sigma. Either a vector of length `K` or a scalar.}

\item{omega}{Baseline scale matrix for Sigma. Either a `q x q` matrix or a `q x q x K` array.}
}
\value{
A list consisting of a vector of length N for \eqn{z}, and a N x K matrix for \eqn{lambda} (see details).
}
\description{
Complete log-likelihood for the normal-normal-mixture model.
}
