% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-sampler.R
\name{relabel}
\alias{relabel}
\title{Labels for recovering cluster order according to the ground truth mu.}
\usage{
relabel(true_mu, sample_mu, K)
}
\arguments{
\item{true_mu}{Ground truth mu of dimension `K x q`. Can be cluster means estimated with known membership of each observation.}

\item{sample_mu}{MCMC samples of mu of dimension `K x q x iteration` from [mbc::gibbs_sampler()].}

\item{K}{Number of clusters (positive integer).}
}
\value{
A list consisting `order_param` (ordering for parameters) and `order_z` (ordering for z), both of length `K`.
}
\description{
Labels for recovering cluster order according to the ground truth mu.
}
\details{
The original order of the shuffled cluster labels is estimated with the ground truth mu and MCMC sample mu, by sequentially finding the minimum difference between the two.
}
