\name{rot.random}
\alias{rot.random}
\title{
Random orthogonal rotation
}
\description{
Generate a \code{k}-dimensional random orthogonal rotation matrix.
}
\usage{
rot.random(k)
}
\arguments{
  \item{k}{
      the number of dimensions.
  }
}

\references{
Mezzadri, F. 2007. How to generate random matrices from the classical compact groups,
Notices of the American Mathematical Society, 54:592–604.
}
