% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{solveV}
\alias{solveV}
\title{Solve method for variance matrices.}
\usage{
solveV(V, x, ldV = FALSE)
}
\arguments{
\item{V}{Variance matrix.}

\item{x}{Optional vector or matrix for which to solve system of equations.  If missing calculates inverse matrix.}

\item{ldV}{Optionally compute log determinant as well.}
}
\value{
Matrix solving system of equations and optionally the log-determinant.
}
\description{
Solve method for variance matrices.
}
\details{
This function is faster and more stable than `base::solve()` when `V` is known to be positive-definite.
}
\author{
Martin Lysy
}
