% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_mu}
\alias{update_mu}
\title{Conditional update for \eqn{\mu}.}
\usage{
update_mu(z, mu, Theta, Sigma, N, q, K, pars = FALSE)
}
\arguments{
\item{z}{Cluster membership indicator vector of length `N`.}

\item{mu}{Original cluster means matrix of dimension `K x q`.}

\item{Theta}{Unobserved data matrix of dimension `N x q`.}

\item{Sigma}{Cluster variance array of dimension `q x q x K`.}

\item{N}{Number of observations (positive integer).}

\item{q}{Size of \eqn{\theta_i} (positive integer).}

\item{K}{Number of clusters (positive integer).}

\item{pars}{Return the means and variances.}
}
\value{
A `K x q` matrix for \eqn{\mu} drawn from its conditional posterior distribution.
If pars is TRUE, a list of parameters of the conditional posterior distribution is returned.
}
\description{
Conditional update for \eqn{\mu}.
}
\details{
The conditional posterior distribution of \eqn{\mu} is given by:
```
\eqn{\mu_k | A - \{ \mu_1,...,\mu_K \} \sim N(\bar \theta_k, Sigma_k / N_k)},
\eqn{\sum_{i=1}^N 1[z_i = k]}.
```
Hence `mu` is sampled from the multivariate normal distribution given all other parameters.
Moreover, updating `mu` is allowed for only nonempty clusters (`N_k` > 0).
}
