% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_rho}
\alias{update_rho}
\title{Conditional update for Gaussian Mixture model weight \eqn{\rho}.}
\usage{
update_rho(z, K, pars = FALSE)
}
\arguments{
\item{z}{Cluster membership indicator vector of length `N`.}

\item{K}{Number of clusters (positive integer).}

\item{pars}{Return the vector length `K` of number of observations.}
}
\value{
A vector of length `K` of \eqn{\rho} drawn from its conditional posterior distribution.
If pars is TRUE, a list of parameters of the conditional posterior distribution is returned.
}
\description{
Conditional update for Gaussian Mixture model weight \eqn{\rho}.
}
\details{
The conditional posterior distribution of \eqn{\rho} is given by:
```
\eqn{\rho | A - \{ \rho \} \sim Dirichlet(\alpha)},    \eqn{\alpha = (N_1 + 1 ,..., N_K + 1)}.
```
Hence `rho` is sampled from the Dirichlet distribution given all other parameters.
}
