% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_Sigma}
\alias{update_Sigma}
\title{Conditional update for cluster variance \eqn{\Sigma}.}
\usage{
update_Sigma(z, Theta, mu, q, K, v_k, omega, pars = FALSE)
}
\arguments{
\item{z}{Cluster membership indicator vector of length `N`.}

\item{Theta}{Unobserved data matrix of dimension `N x q`.}

\item{mu}{Cluster means matrix of dimension `K x q`.}

\item{q}{Size of \eqn{\theta_i} (positive integer).}

\item{K}{Number of clusters (positive integer).}

\item{v_k}{Baseline degree of freedom for Sigma. Either a vector of length `K` or a scalar.}

\item{omega}{Baseline scale matrix for Sigma. Either a `q x q` matrix or a `q x q x K` array.}

\item{pars}{Return the scale parameter and degree of freedom.}
}
\value{
A `q x q x K` array of \eqn{\Sigma} drawn from its conditional posterior distribution.
If pars is TRUE, a list of parameters of the conditional posterior distribution is returned.
}
\description{
Conditional update for cluster variance \eqn{\Sigma}.
}
\details{
The conditional posterior distribution of \eqn{\Sigma} is given by:
```
\eqn{\Sigma_k | A - \{ \Sigma_1,...,\Sigma_K \} \sim InvWish(\Omega_k + 1[z_i=k] (\theta_i - \mu_k)(\theta_i - \mu_k)', N_k + v_k)}.
```
Hence `Sigma` is sampled from the Inverse-Wishart distribution given all other parameters.
}
