% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_Theta}
\alias{update_Theta}
\title{Conditional update for the unobserved data \eqn{\Theta}.}
\usage{
update_Theta(y, V, z, mu, Sigma, N, pars = FALSE)
}
\arguments{
\item{y}{Observed data of dimension `N x q`.}

\item{V}{Observed Fisher information array of dimension `q x q x N`.}

\item{z}{Cluster membership indicator vector of length `N`.}

\item{mu}{Cluster means matrix of dimension `K x q`.}

\item{Sigma}{Cluster variance array of dimension `q x q x K`.}

\item{N}{Number of observations (positive integer).}

\item{pars}{Return the means and variances.}
}
\value{
A `N x q` matrix for \eqn{\Theta} drawn from its conditional posterior distribution.
If pars is TRUE, a list of parameters of the conditional posterior distribution is returned.
}
\description{
Conditional update for the unobserved data \eqn{\Theta}.
}
\details{
The conditional posterior distribution of \eqn{\Theta} is given by:
```
\eqn{\theta_i | A - \{ \theta_i \} \sim N(G_i(y_i - \mu_{z_i}) + \mu_{z_i}, G_i V_i)},
\eqn{G_i = \Sigma_{z_i} (V_i + \Sigma_{z_i})^{-1}}.
```
Hence `Theta` is sampled from the random-effects normal distribution given all other parameters.
}
