% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_z}
\alias{update_z}
\title{Conditional update for cluster membership indicator \eqn{z}.}
\usage{
update_z(Sigma, Theta, mu, rho, N, q, K)
}
\arguments{
\item{Sigma}{Cluster variance array of dimension `q x q x K`.}

\item{Theta}{Unobserved data matrix of dimension `N x q.`}

\item{mu}{Cluster means matrix of dimension `K x q`.}

\item{rho}{Gaussian Mixture model weight of vector of length `K`.}

\item{N}{Number of observations (positive integer).}

\item{q}{Size of \eqn{\theta_i} (positive integer).}

\item{K}{Number of clusters (positive integer).}
}
\value{
A list consisting of a vector of length `N` for \eqn{z}, and a `N x K` matrix for \eqn{lambda} (see details).
}
\description{
Conditional update for cluster membership indicator \eqn{z}.
}
\details{
The conditional posterior distribution of \eqn{z} is given by:
```
\eqn{\kappa_{ik} = log \rho_k + log \phi(\theta_i | \mu_k, \Sigma_k) + CONST},
\eqn{z_i | A - \{ z \} \sim Multinomial(K, \lambda_i)},    \eqn{\lambda_{ik} = exp(\kappa_{ik}) / \sum^K_{m=1} exp(\kappa_{im}))}.
```
Hence, `z` is sampled from the Multinomial distribution given all other parameters.
}
