% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc-package.R
\docType{package}
\name{mbc-package}
\alias{mbc}
\alias{mbc-package}
\title{mbc: Model-Based Classification}
\description{
Classification method using Gibbs sampler with a hierarchical Bayesian clustering model.
}
\keyword{internal}
