% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBC-QDM.R
\name{rot.random}
\alias{rot.random}
\title{Random orthogonal rotation}
\usage{
rot.random(k)
}
\arguments{
\item{k}{the number of dimensions.}
}
\description{
Generate a \code{k}-dimensional random orthogonal rotation matrix.
}
\references{
Mezzadri, F. 2007. How to generate random matrices from the
classical compact groups, Notices of the American Mathematical Society,
54:592–604.
}
